Add Crimea to teleport locations
To those who are willing to dig this far to figure out how
Mt Detect got it's name: Me and Bombita decided it
spur-of-the moment at 1:30 AM on a sunday night.
It's the only mountain that has alpha transparency.

I leave you with a bit of airmash holy scripture:

Looking down from is post in Greenland,
Detect said "Let there be a flag."
Thus he spawned the blue flag,
And it was good.
Then Detect said: "Let not the blue flag be alone"
So he took a thread from the blue flag,
and from that thread, he created the red flag.
Upon his back, he carried the red flag to its base.
And it was good.
Thus was the birth of CTF.

Amen

The book of Detect 53:14