Fix edge case within upgrade visibility
Previously, if an upgrade was spawned and moved out of the player's view
radius within the same frame, then the player would never get an
EventLeaveHorizon packet and the upgrade would remain a ghost forever.

This commit adds a test case and fixes the issue by explicitly adding
the upgrade to the relvant visibility sets when it is spawned.