Move buffer copies off of the game thread.
The game thread spends a lot of time cloning buffers containing
serialized messages to send to clients. However, while there is only one
game thread there are many client threads so this commit shifts the
cloning work to the client threads to prevent lag when there are large
numbers of clients on the server (around 400).

It also includes a bundle to reuse hashmap capacity for visibility
updates as one thing that showed up a lot in profiling was resizing
hashmaps within the visibility subsystem.