Fix status page to actually return the number of players
Due to some changes in tungstenite or tokio-tungstenite it is no longer
possible to return a response from accept_hdr_async that is not either
1. the continuation of the websocket handshake, or
2. an error response.

This causes a problem for airmash servers since the client actually
queries the server endpoint and expects a 200 OK response with a json
payload if the request is not a websocket request. To work around this,
I have manually reimplemented the websocket handshake so that we can
return the expected results when a non-websocket request comes in.