Make picking up a powerup grant the corresponding effect (#155)
Currently, if a powerup is spawned and a player goes to pick it up then the powerup will despawn but the player will not be given the corresponding effect. This PR changes that to now give the player the expected effect. I've also included tests to verify that everything works as expected.