Make mohawk missiles fire on alternating sides of the plane (#159)
Currently, all missiles from all planes are fired from the centre of the plane. However, this is not correct for the mohawk which is supposed to fire with the offset switching sides with each shot. This PR adds this behaviour. Note that this involves a change to the config so I've added a small utility to export the default config and then gone and updated it.