Convert uses of `PlaneType` within server to instead use plane prototypes (#198)
This PR converts pretty much all uses of `PlaneType` and `Config::planes` to instead use plane prototypes. Along with that, it also converts a bunch of missile-related code to use missile prototypes.

It passes all tests, although due to the fact that there are multiple different configs it breaks custom configs. This will get fixed once the conversion is completed and the new config system is being used.