Sequence `PLAYER_TYPE` messages before `PLAYER_RESPAWN` messages for plane change commands (#202)
The issue motivating this commit was that Q and E would fail to do anything when playing as a mohawk. The root cause of this issue was due to a bit of a desync between what the server thinks the player's type is and what the client thinks the player's type is. This is because the type of the current player is stored in two places in the client:
1. `Players.getMe().type`
2. `game.myType`

`Players.getMe().type` is updated when the client receives a `PLAYER_TYPE` packet for the current player. However, `game.myType` is only updated when the player respawns. This isn't usually noticeable because the _only_ thing that `game.myType` is used for is determining whether Q and E will trigger strafing or do nothing.

The fix, as contained in this commit, is to make sure to send the `PLAYER_TYPE` packet before sending the `PLAYER_RESPAWN` packet. This ensures that the client has the right type before it gets updated in `Player.respawn`.

Fixes #201