Ensure that structs with options deserialize properly (#231)
The problem is that for lua tables there is no difference between a key
that doesn't exist and one that corresponds to nil (in fact assigning
nil is how you remove a key from the table). However, we serialize
options to nil when they are None so we can run into this with structs
pretty easily.

The solution here is to track the extra fields in the struct and
explicitly deserialize a nil as their value if they are not found within
the table that is being deserialized. This ensures that all the
downstream structs in server-config deserialize properly.

We will still have some issues around maps that contain options but
that's a much rarer case than structs with options so I'm ok with that
for now.